update: auto select first item in `DocmentsListView`.
Seems like `ListView.clear` and `ListView.extend` are async methods.
There seems to be some internal mechanism for synchronization, so they
didn't need to be awaited before, but now, when I need to select (set
the index) after adding elements, it sometimes would just not work
unless I explicitly await the `extend` and then set the index only after
it's done adding the elements.
This change this makes the watcher async as well.